Add development init data. Router, controller, query setup work in progress. Commit time used 6 h 30 min, total 20,5

+++ server/models/item.ts
@@ -43,17 +43,17 @@ Item.init({
     allowNull: false
   },
   unitSize: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL,
     allowNull: false
   },
   brand: {
     type: DataTypes.TEXT
   },
   price: {
-    type: DataTypes.INTEGER
+    type: DataTypes.DECIMAL
   },
   pricePerUnit: {
-    type: DataTypes.INTEGER
+    type: DataTypes.DECIMAL
   },
 
 }, {
